clean: Instantiate the aws client once
